re-organize the items files
