+++ src/ui/MessageMenu/index.tsx
@@ -1,3 +1,2 @@
-export { MessageMenu, type MessageMenuProps } from './messageMenu';
-export { MessageMenuProvider, useMessageMenuContext, type MessageMenuContextProps } from './messageMenuContext';
-export { MobileMessageMenu } from './mobileMessageMenu';
+export { MessageMenu, type MessageMenuProps } from './MessageMenu';
+export { MessageMenuProvider, useMessageMenuContext, type MessageMenuContextProps } from './MessageMenuProvider';
