+++ src/ui/MessageMenu/menuItems/mobileItems.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import { FileMessage } from '@sendbird/chat/message';
 
-import { MenuItem, PrebuildMenuItemPropsType } from '../MessageMenu/items';
-import { copyToClipboard, isUserMessage, isFailedMessage } from '../../utils';
-import { useLocalization } from '../../lib/LocalizationContext';
-// import { useMessageMenuContext } from './MessageMenuContext';
-import Icon, { IconTypes, IconColors } from '../Icon';
-import Label, { LabelColors, LabelTypography } from '../Label';
-import { MobileMessageMenuContextProps, useMessageMenuContext } from './messageMenuContext';
+import { MenuItem, PrebuildMenuItemPropsType } from './menuItems';
+import { copyToClipboard, isUserMessage, isFailedMessage } from '../../../utils';
+import { useLocalization } from '../../../lib/LocalizationContext';
+import Icon, { IconTypes, IconColors } from '../../Icon';
+import Label, { LabelColors, LabelTypography } from '../../Label';
+import { MobileMessageMenuContextProps, useMessageMenuContext } from '../MessageMenuProvider';
 
 export const CopyMenuItem = (props: PrebuildMenuItemPropsType) => {
   const { stringSet } = useLocalization();
