+++ src/ui/MobileMenu/bottomSheetMenuItem.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MenuItemProps } from '../MessageMenu/items';
+import { MenuItemProps } from '../MessageMenu/menuItems/menuItems';
 import { classnames } from '../../utils/utils';
 
 export const BottomSheetMenuItem = ({
